backend2/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
google-generativeai
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers